requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiodns==3.2.0
    # via
    #   -r requirements.in
    #   ccxt
aiofiles==24.1.0
    # via
    #   -r requirements.in
    #   ccxt
    #   coinbase-advanced-py
    #   pandas
    #   sqlalchemy
    #   tabulate
    #   yarl
    #   tzdata
    #   python-dotenv
    #   requests
    #   aiosqlite
    #   aiodns
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.10.11
    # via
    #   -r requirements.in
    #   ccxt
aiosignal==1.3.1
    # via
    #   -r requirements.in
    #   aiohttp
aiosqlite==0.20.0
    # via -r requirements.in
async-timeout==5.0.1
    # via
    #   -r requirements.in
    #   aiohttp
attrs==24.2.0
    # via
    #   -r requirements.in
    #   aiohttp
backoff==2.2.1
    # via coinbase-advanced-py
bottleneck==1.4.2
    # via -r requirements.in
ccxt==4.4.61
    # via -r requirements.in
certifi==2024.8.30
    # via
    #   -r requirements.in
    #   ccxt
    #   requests
cffi==1.17.1
    # via
    #   -r requirements.in
    #   cryptography
    #   pycares
charset-normalizer==3.4.0
    # via
    #   -r requirements.in
    #   requests
coinbase==2.1.0
    # via -r requirements.in
coinbase-advanced-py==1.8.2
    # via -r requirements.in
cryptography==44.0.0
    # via
    #   -r requirements.in
    #   ccxt
    #   coinbase-advanced-py
frozenlist==1.5.0
    # via
    #   -r requirements.in
    #   aiohttp
    #   aiosignal
greenlet==3.1.1
    # via
    #   -r requirements.in
    #   sqlalchemy
idna==3.10
    # via
    #   -r requirements.in
    #   requests
    #   yarl
multidict==6.1.0
    # via
    #   -r requirements.in
    #   aiohttp
    #   yarl
numexpr==2.10.2
    # via -r requirements.in
numpy==2.1.3
    # via
    #   -r requirements.in
    #   bottleneck
    #   numexpr
    #   pandas
pandas==2.2.3
    # via -r requirements.in
propcache==0.2.1
    # via yarl
pycares==4.5.0
    # via
    #   -r requirements.in
    #   aiodns
pycparser==2.22
    # via
    #   -r requirements.in
    #   cffi
pycryptodome==3.21.0
    # via coinbase
pyjwt==2.10.1
    # via coinbase-advanced-py
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.2
    # via
    #   -r requirements.in
    #   pandas
requests==2.32.3
    # via
    #   -r requirements.in
    #   ccxt
    #   coinbase
    #   coinbase-advanced-py
six==1.16.0
    # via
    #   -r requirements.in
    #   coinbase
    #   python-dateutil
sqlalchemy==2.0.36
    # via -r requirements.in
tabulate==0.9.0
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   aiosqlite
    #   ccxt
    #   multidict
    #   sqlalchemy
tzdata==2024.2
    # via
    #   -r requirements.in
    #   pandas
urllib3==2.2.3
    # via
    #   -r requirements.in
    #   requests
websockets==13.1
    # via
    #   -r requirements.in
    #   coinbase-advanced-py
yarl==1.18.0
    # via
    #   -r requirements.in
    #   aiohttp
    #   ccxt

# The following packages are considered to be unsafe in a requirements file:
# setuptools
